Added google auth protection layer to the app
Added an UserAuthContext to check if user is authenticated.
Added a sign out button in the forum page to sign out the current user.
forum page is blocked if the user doesnt sign in
Everything is done in the client side.
Possible TODO:
Integrate the phone sign in with firebase (has the lowest priority)